requirements.txt
torch>=1.10.0
numpy>=1.20.0
tqdm>=4.62.0
pillow>=8.3.0
gradio>=3.0.0
wandb>=0.12.0
torchvision>=0.11.0
setuptools>=58.0.0
wheel>=0.37.0
